+++ persistent-test/src/TreeTest.hs
@@ -43,27 +43,27 @@ specsWith runDb = describe "tree" $ do
     describe "entityDef" $ do
         let EntityDef{..} = entityDef (Proxy :: Proxy Tree)
         it "has the right haskell name" $ do
-            entityHaskell `shouldBe` HaskellName "Tree"
+            entityHaskell `shouldBe` EntityNameHS "Tree"
         it "has the right DB name" $ do
-            entityDB `shouldBe` DBName "trees"
+            entityDB `shouldBe` EntityNameDB "trees"
 
     describe "foreign ref" $ do
         let [ForeignDef{..}] = entityForeigns (entityDef (Proxy :: Proxy Tree))
         it "has the right haskell name" $ do
             foreignRefTableHaskell `shouldBe`
-                HaskellName "Tree"
+                EntityNameHS "Tree"
         it "has the right db name" $ do
             foreignRefTableDBName `shouldBe`
-                DBName "trees"
+                EntityNameDB "trees"
         it "has the right constraint name" $ do
             foreignConstraintNameHaskell `shouldBe`
-                HaskellName "fkparent"
+                ConstraintNameHS "fkparent"
         it "has the right DB constraint name" $ do
             foreignConstraintNameDBName `shouldBe`
-                DBName "treesfkparent"
+                ConstraintNameDB "treesfkparent"
         it "has the right fields" $ do
             foreignFields `shouldBe`
-                [ ( (HaskellName "parent", DBName "parent")
-                  , (HaskellName "name", DBName "name")
+                [ ( (FieldNameHS "parent", FieldNameDB "parent")
+                  , (FieldNameHS "name", FieldNameDB "name")
                   )
                 ]