+++ persistent/ChangeLog.md
@@ -1,12 +1,15 @@
 # Changelog for persistent
 
-## 2.11.1
+## 2.12
 
+* [#1162](https://github.com/yesodweb/persistent/pull/1162)
+  * Replace `askLogFunc` with `askLoggerIO`
+* Decomposed `HaskellName` into `ConstraintNameHS`, `EntityNameHS`, `FieldNameHS`. Decomposed `DBName` into `ConstraintNameDB`, `EntityNameDB`, `FieldNameDB` respectively. [#1174](https://github.com/yesodweb/persistent/pull/1174)
+* Use `resourcet-pool` to break out some `Data.Pool` logic [#1163](https://github.com/yesodweb/persistent/pull/1163)
 * [#1178](https://github.com/yesodweb/persistent/pull/1178)
   * Added 'withBaseBackend', 'withCompatible' to ease use of base/compatible backend queries in external code.
 
 ## 2.11.0.2
-
 * Fix a bug where an empty entity definition would break parsing of `EntityDef`s. [#1176](https://github.com/yesodweb/persistent/issues/1176)
 
 ## 2.11.0.1